refactor: update h3
