+++ playground/routes/route.ts
@@ -1,5 +1,5 @@
-import { defineHandler } from "h3";
-
-export default defineHandler((event) => {
-  return "Response from routes/route.ts";
-});
+export default {
+  fetch() {
+    return new Response("Response from routes/routes.ts!");
+  },
+};
