+++ playground/server.ts
@@ -1,12 +1,9 @@
 export default {
-  async fetch(req: Request) {
+  async fetch(req: Request): Promise<Response | void> {
     console.log(`[${req.method}] ${req.url}`);
     const url = new URL(req.url);
-    if (url.pathname === "/server") {
-      return new Response("Response from server.ts");
-    }
     if (url.pathname === "/") {
-      return fetch(req, { viteEnv: "ssr" } as any);
+      return new Response("Response from server.ts");
     }
   },
 };
