+++ src/presets/_nitro/nitro-dev.ts
@@ -1,4 +1,6 @@
+import { runtimeDir } from "nitro/runtime/meta";
 import { defineNitroPreset } from "../_utils/preset";
+import { join } from "pathe";
 
 const nitroDev = defineNitroPreset(
   {
@@ -8,6 +10,12 @@ const nitroDev = defineNitroPreset(
       serverDir: "{{ buildDir }}/dev",
       publicDir: "{{ buildDir }}/dev",
     },
+    handlers: [
+      {
+        route: "/_nitro/tasks/**",
+        handler: join(runtimeDir, "internal/routes/dev-tasks"),
+      },
+    ],
     externals: { trace: false },
     serveStatic: true,
     inlineDynamicImports: true, // externals plugin limitation
